requirements.txt
pillow
tqdm
pytorch-msssim
tensorboard
tensorboardx